Add more of the record structure for FO4
This fleshes out some more of the record structure for FO4.

In particular it adds support for the `TRNS`, `DMGT`, `CMPO` records, and expands the `COBJ` records to reflect their state in FO4.
This also includes a couple of other fixes due to `PTRN` subrecords being added in many places. There are still quite a few places that need the `PTRN` subrecords added though.
This also handles the extra 8-bytes that are at the start of the `DMDT` and `MODT` subrecords in the `DESTRUCT` and `MODEL` groups respectively.
This fixes most places so that the type is read correctly, however it appears that the extra 8-bytes require a condition that I haven't looked into yet. As the majority of places have the extra 8-bytes, I've left the change in this PR.